+++ src/components/FolderPage/FolderList.js
@@ -7,7 +7,7 @@ import UtilIcons from "./UtilIcons.js";
 import add from "../../assets/add.svg";
 import FolderButtons from "./FolderButtons";
 import useModal from "../../hooks/useModal";
-import ModalAdd from "../Modal/ModalAdd";
+import ModalAddFolder from "../Modal/ModalAddFolder.js";
 
 const FolderList = () => {
   const folderList = useFolderList();
@@ -55,7 +55,7 @@ const FolderList = () => {
           폴더 추가
           <img src={add} alt="add icon" />
         </button>
-        <ModalAdd isOpen={showModal} onClose={handleCloseModal} />
+        <ModalAddFolder isOpen={showModal} onClose={handleCloseModal} />
       </div>
       <div className="selectedFolderName">
         {selectedFolderName}
