+++ src/components/FolderPage/UtilIcons.css
@@ -7,4 +7,5 @@
 .utilContainer img {
   position: relative;
   top: 4px;
+  cursor: pointer;
 }
