+++ src/components/Modal/ModalAddFolder.js
@@ -1,20 +1,16 @@
 import ModalContainer from "./ModalContainer";
 
 const ModalAddFolder = ({ isOpen, onClose }) => {
-  const addFolderContent = (
+  const addContent = (
     <>
-      <h2>폴더에 추가 </h2>
+      <h2>폴더 추가</h2>
       <input placeholder="내용 입력" className="modalInput"></input>
-      <button>추가하기</button>
+      <button className="blueButton">추가하기</button>
     </>
   );
 
   return (
-    <ModalContainer
-      isOpen={isOpen}
-      onClose={onClose}
-      children={addFolderContent}
-    />
+    <ModalContainer isOpen={isOpen} onClose={onClose} children={addContent} />
   );
 };
 export default ModalAddFolder;
